requirements.txt
lingpy~=2.6.13
lingrex~=1.4.2
lexibase
sentence-transformers>=2.2.2
torch>=2.1.0
pandas>=2.1.0
openpyxl>=3.1.2
segments>=2.3.0
